### 2016-11-13 #109 @janschulz
Looked at this and I found that the install instruction contain this:

> Survival Volume is most fully featured with mpld3 version 0.3dev post commit 559dfc6. Currently the easiest way to install a compatible version is to install from the current development master branch on github.

On mpl3d, there is an issue about releasing a 0.3 version (https://github.com/mpld3/mpld3/issues/398), so I would currently wait until that has happend
