### 2016-11-13 #109 @arfon
@janschulz - given your comment above it sounds like you're willing to review this for us? If so, I'll get started with opening the actual review issue.
