### 2016-11-13 #109 @genomematt
Hi Jan,
The timeline for mpld3 v0.3 is still uncertain, and I have another publication coming in which I wish to use (and therefore cite) this package.
The only significant issue with mpld3 v0.2 is the rendering of the steps in the Kaplan Meier graph.  This only impacts the dual plots and only when done as interactive html.
The html version of the user guide was constructed with mpld3 v0.2

There are some minor compatibility issues between mpld3 v0.2 and matplotlib >1.3 but none of them impact the core functionality.

Also I just noticed while confirming all the above that you need to do pip3 install —pre because I marked it as a beta.
I will roll it over to 1.0.0 fixing this when I address any review points and mint a new release for the final paper version.

Regs,
Matt

> On 14 Nov 2016, at 1:23 AM, Jan Schulz <notifications@github.com> wrote:
>
> Looked at this and I found that the install instruction contain this:
>
> Survival Volume is most fully featured with mpld3 version 0.3dev post commit 559dfc6. Currently the easiest way to install a compatible version is to install from the current development master branch on github.
>
> On mpl3d, there is an issue about releasing a 0.3 version (mpld3/mpld3#398 <https://github.com/mpld3/mpld3/issues/398>), so I would currently wait until that has happend
>
