### 2016-11-14 #109 @arfon
@janschulz @genomematt let's continue the review over in #111
