### 2016-11-16 #111 @arfon
@janschulz - I've just added the compiled paper PDF for your review.
