### 2016-12-08 genomematt/survivalvolume#2 @janschulz
Docs on parse are garbled
https://genomematt.github.io/survivalvolume/docs/parse.m.html

-> The "Argument" and "Return" sections do not stand out like they do in the [plot](https://genomematt.github.io/survivalvolume/docs/plot.m.html) section.
