### 2016-12-08 #111 @janschulz
Create a new conda env with `conda create -n "survivalvolume" python=3.5 matplotlib mpld3 lifelines scipy pandas numpy xlrd ipykernel` -> worked like a charm :-)

* License is GPL V3 -> :heavy_check_mark:
* Version in github and pip is v1.0.0b1 :question:
* most commits come from @genomematt :heavy_check_mark:
* installation works as outlines (ok, had to use `pip` instead of `pip3`, but both pypi and git works)
* Functionality: I went through the ipynb and the plots showed up as described. :heavy_check_mark:
* Performance: no performance claims were maid but with the ipynb with example data, there was no performance problem. :heavy_check_mark:

* Statement of need: "A set of tools for parsing measurement with threshold over time data (eg tyre wear, tumour treatment studies) and generating interactive and static plots". as a not-from-that-field guy, I have actually no idea what I can do with that, but I guess if you working in that field you do -> a sentence of who the target audience is would have helped me. The description in the paper is actually longer and and I actually understood it a bit :-) -> I give the paper version a "passed" because my feeling is that it contains all the buzzwords needed to get the attention of people in the field and I got the impression that I can't use it to do a customers survival analysis (If I could, please talk to me :-)) but please copy that to the real docs as well :question:
* Api Docs: https://genomematt.github.io/survivalvolume/docs/parse.m.html has garbled parameter docs (=they do not show up in bold like in https://genomematt.github.io/survivalvolume/docs/plot.m.html) -> https://github.com/genomematt/survivalvolume/issues/2, otherwise :heavy_check_mark:
* travis shows 18 tests. Test coverage seems ok, some tests even check matplotlib internal stuff (you might like [pytest-mpl](https://pypi.python.org/pypi/pytest-mpl) to test plots directly) :heavy_check_mark:
* guidelines: short and to the point :heavy_check_mark:

# Paper
* authors have OrcIDs :+1: and affiliations :heavy_check_mark:
* statement of need: see above: :heavy_check_mark:
* DOIs are shown where I think they are available

# open
- verison: 1.0.0 vs beta?
- It would be nice if the readme doc gets the "statement of need" from the paper
