### 2016-12-11 #111 @genomematt
Thanks @janschulz I will fix those documentation issues, and mint a final 1.0.0 version (that was the purpose of being still a beta release).
