### 2016-12-11 f80940f
fix parse.py docstrings
