### 2016-12-12 c1cba47
update version strings
