+++ survivalvolume/tests/test_all.py
@@ -23,7 +23,7 @@ __author__ = "Matthew Wakefield"
 __copyright__ = "Copyright 2011-2016 Matthew Wakefield, The Walter and Eliza Hall Institute and The University of Melbourne"
 __credits__ = ["Matthew Wakefield",]
 __license__ = "GPLv3"
-__version__ = "1.0.0b1"
+__version__ = "1.0.0"
 __maintainer__ = "Matthew Wakefield"
 __email__ = "wakefield@wehi.edu.au"
 __status__ = "Beta"
