### 2016-12-12 #111 @genomematt
Ok @janschulz genomematt@c1cba47e490f766affffc2f108915a57bd5b0b87 should fix all the outstanding issues.
Uploaded to PyPI and new Zenodo DOI is 10.5281/zenodo.200268
