### 2016-12-12 #111 @genomematt
As far as retention rate analysis / customer survival analysis there is only one potential way of parameterising that would make sense for survivalvolume.  If you had an activity measure you could use this in place of volume.  The obvious one would be days (or minutes) of inactivity, and then at a threshold you would declare them inactive.  Probably better with cumulative days of inactivity so it increases monotonically.  You could then look at patterns of inactivity in the activity graph, mouse over a specific customer etc.
