### 2016-12-12 #111 @arfon
@janschulz can you confirm we're good to accept here?
