### 2016-12-12 #111 @janschulz
@arfon: yes, i looked at the new readme and the version and bot LGTM :-)

@genomematt Thanks for the timely response, especially after I let this slide!  :-/
