### 2016-12-12 #111 @janschulz
@genomematt Ah, and also thank you for your explanation on using it for customer retention analysis!
